+++ demo/demo_form.php
@@ -29,6 +29,7 @@ $form_def=array(
     'max_length' => 10,
     'html_attributes'   =>array("style"=>"border: 2px solid black;"),
     'check'	=> array("fun", array("js"=>"name_check", "php"=>"name_check")),
+    'non_used_value' => 'FOO',
   ),
   'nickname'   =>array(
     'name'      =>'Nickname',